refactor: moved unit.conftest to src/conftest
so that plugin tests can access the unit test fixtures
