--- docsrc/modules/test.rst
@@ -1,8 +1,0 @@
-test
-===============
-
-.. toctree::
-   :maxdepth: 4
-
-   test.unit
-
