--- docsrc/modules/test.unit.conftest.rst
@@ -1,6 +1,0 @@
-conftest
-========
-.. automodule:: test.unit.conftest
-   :members:
-   :undoc-members:
-   :show-inheritance:
