--- docsrc/modules/test.unit.rst
@@ -1,8 +1,0 @@
-test.unit
-===============
-
-.. toctree::
-   :maxdepth: 4
-
-   test.unit.conftest
-
