+++ src/test/unit/web_interface/test_app_comparison_text_files.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import pytest
 
+from conftest import CommonIntercomMock, WebInterfaceUnitTestConfig
 from test.common_helper import TEST_TEXT_FILE, TEST_TEXT_FILE2, CommonDatabaseMock, create_test_firmware
-from test.unit.conftest import CommonIntercomMock, WebInterfaceUnitTestConfig
 
 
 class MockInterCom(CommonIntercomMock):
